refactor: flatten workspace — reef is now the root crate
The repo is named `reef` and ships a single binary — nesting that binary
at `crates/reef-host/` (or even `crates/reef/`) was one layer of
indirection past what the project actually is.

- Root `Cargo.toml` is now `[package] name = "reef"`, with a `[workspace]`
  section that just lists `crates/test-support` (the one dev-only helper
  crate). Root also owns `[[bin]] name = "reef"`, all `[dependencies]` /
  `[dev-dependencies]`, and both `[[bench]]` entries. rust-analyzer and
  `cargo doc` now show paths as `reef::git::graph::LaneCell` instead of
  `reef_host::git::graph::LaneCell`.
- `src/` / `tests/` / `benches/` moved to the repo root (previously under
  `crates/reef-host/`).
- `crates/reef-host/` directory deleted.
- Every `use reef_host::...` rewritten to `use reef::...` across src,
  tests, benches, fuzz, and the skills docs.
- `fuzz/Cargo.toml`: dep path `../crates/reef-host` → `..` (the root).
- `.github/workflows/release.yml`: `-p reef-host` → `-p reef`.
- `skills/testing-reef/*.md`: `crates/reef-host/**` paths rewritten to
  root-relative (`tests/...`, `src/...`, `benches/...`).

No code changes, just mv + rename. 124/124 tests green, clippy + fmt
clean.
