--- crates/reef-host/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "reef-host"
-version = "0.1.0"
-edition = "2024"
-publish = false
-
-[lints]
-workspace = true
-
-[[bin]]
-name = "reef"
-path = "src/main.rs"
-
-[dependencies]
-crossterm     = "0.29.0"
-git2          = { version = "0.20.4", default-features = false }
-ratatui       = "0.30.0"
-similar       = "3.1.0"
-unicode-width = "0.2.2"
-syntect       = { version = "5.2", default-features = false, features = ["default-fancy"] }
-two-face      = { version = "0.5", default-features = false, features = ["syntect-fancy"] }
-notify        = "6"
-ignore        = "0.4"
-serde_json    = { workspace = true }
-
-[dev-dependencies]
-tempfile     = "3"
-insta        = { version = "1", features = ["filters"] }
-criterion    = { version = "0.5", features = ["html_reports"] }
-proptest     = "1"
-test-support = { path = "../test-support" }
-
-[[bench]]
-name = "highlight"
-harness = false
-
-[[bench]]
-name = "graph"
-harness = false
