+++ src/lib.rs
@@ -1,4 +1,4 @@
-//! Library crate for reef-host — exposes modules as public API so that
+//! Library crate for reef — exposes modules as public API so that
 //! integration tests under `tests/` can exercise them. The `reef` binary
 //! (src/main.rs) is a thin wrapper that consumes this library.
