+++ src/ui/commit_detail_panel.rs
