+++ src/ui/diff_panel.rs
