+++ src/ui/file_preview_panel.rs
